refactor: moving typeCommand to lib as it is a very special command
